スコア詳細のリファクタリング
ロジックとビューの分離
UIのコンポーネント化
